# pkg/api/api.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.6.0 DO NOT EDIT.
package api

import (
	"context"
	"encoding/json"
	"fmt"
	"net/http"
	"time"

	"github.com/labstack/echo/v4"
	"github.com/oapi-codegen/runtime"
	strictecho "github.com/oapi-codegen/runtime/strictmiddleware/echo"
)

// Defines values for NamespaceStatus.
const (
	Active      NamespaceStatus = "active"
	Terminating NamespaceStatus = "terminating"
)

// Valid indicates whether the value is a known member of the NamespaceStatus enum.
func (e NamespaceStatus) Valid() bool {
	switch e {
	case Active:
		return true
	case Terminating:
		return true
	default:
		return false
	}
}

// Error defines model for Error.
type Error struct {
	// Code HTTP статус код
	Code int `json:"code"`

	// Description Дополнительная информация об ошибке
	Description *string `json:"description,omitempty"`

	// Message Описание ошибки
	Message string `json:"message"`
}

// Namespace defines model for Namespace.
type Namespace struct {
	// CreatedAt Дата и время создания
	CreatedAt *time.Time `json:"created_at,omitempty"`

	// Labels Метки для организации неймспейсов
	Labels map[string]string `json:"labels"`

	// Name Уникальное имя неймспейса
	Name string `json:"name"`

	// Status Статус неймспейса
	Status *NamespaceStatus `json:"status,omitempty"`

	// UpdatedAt Дата и время последнего обновления
	UpdatedAt *time.Time `json:"updated_at,omitempty"`
}

// NamespaceStatus Статус неймспейса
type NamespaceStatus string

// NamespaceCreate defines model for NamespaceCreate.
type NamespaceCreate struct {
	// Labels Метки для организации неймспейсов
	Labels map[string]string `json:"labels"`

	// Name Уникальное имя неймспейса
	Name string `json:"name"`
}

// NamespaceList defines model for NamespaceList.
type NamespaceList struct {
	Items *[]Namespace `json:"items,omitempty"`
}

// NamespaceUpdate defines model for NamespaceUpdate.
type NamespaceUpdate struct {
	// Description Описание неймспейса
	Description *string `json:"description,omitempty"`

	// Labels Метки для организации неймспейсов
	Labels *map[string]string `json:"labels,omitempty"`
}

// NamespaceNameParam defines model for NamespaceNameParam.
type NamespaceNameParam = string

// BadRequestError defines model for BadRequestError.
type BadRequestError = Error

// NotFoundError defines model for NotFoundError.
type NotFoundError = Error

// CreateNamespaceJSONRequestBody defines body for CreateNamespace for application/json ContentType.
type CreateNamespaceJSONRequestBody = NamespaceCreate

// UpdateNamespaceJSONRequestBody defines body for UpdateNamespace for application/json ContentType.
type UpdateNamespaceJSONRequestBody = NamespaceUpdate

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Получить список неймспейсов
	// (GET /namespaces)
	ListNamespaces(ctx echo.Context) error
	// Создать новый неймспейс
	// (POST /namespaces)
	CreateNamespace(ctx echo.Context) error
	// Удалить неймспейс
	// (DELETE /namespaces/{namespaceName})
	DeleteNamespace(ctx echo.Context, namespaceName NamespaceNameParam) error
	// Получить неймспейс по его имени
	// (GET /namespaces/{namespaceName})
	GetNamespace(ctx echo.Context, namespaceName NamespaceNameParam) error
	// Обновить неймспейс
	// (PUT /namespaces/{namespaceName})
	UpdateNamespace(ctx echo.Context, namespaceName NamespaceNameParam) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// ListNamespaces converts echo context to params.
func (w *ServerInterfaceWrapper) ListNamespaces(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.ListNamespaces(ctx)
	return err
}

// CreateNamespace converts echo context to params.
func (w *ServerInterfaceWrapper) CreateNamespace(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.CreateNamespace(ctx)
	return err
}

// DeleteNamespace converts echo context to params.
func (w *ServerInterfaceWrapper) DeleteNamespace(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "namespaceName" -------------
	var namespaceName NamespaceNameParam

	err = runtime.BindStyledParameterWithOptions("simple", "namespaceName", ctx.Param("namespaceName"), &namespaceName, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter namespaceName: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.DeleteNamespace(ctx, namespaceName)
	return err
}

// GetNamespace converts echo context to params.
func (w *ServerInterfaceWrapper) GetNamespace(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "namespaceName" -------------
	var namespaceName NamespaceNameParam

	err = runtime.BindStyledParameterWithOptions("simple", "namespaceName", ctx.Param("namespaceName"), &namespaceName, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter namespaceName: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetNamespace(ctx, namespaceName)
	return err
}

// UpdateNamespace converts echo context to params.
func (w *ServerInterfaceWrapper) UpdateNamespace(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "namespaceName" -------------
	var namespaceName NamespaceNameParam

	err = runtime.BindStyledParameterWithOptions("simple", "namespaceName", ctx.Param("namespaceName"), &namespaceName, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter namespaceName: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.UpdateNamespace(ctx, namespaceName)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithBaseURL(router, si, "")
}

// Registers handlers, and prepends BaseURL to the paths, so that the paths
// can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.GET(baseURL+"/namespaces", wrapper.ListNamespaces)
	router.POST(baseURL+"/namespaces", wrapper.CreateNamespace)
	router.DELETE(baseURL+"/namespaces/:namespaceName", wrapper.DeleteNamespace)
	router.GET(baseURL+"/namespaces/:namespaceName", wrapper.GetNamespace)
	router.PUT(baseURL+"/namespaces/:namespaceName", wrapper.UpdateNamespace)

}

type BadRequestErrorJSONResponse Error

type NotFoundErrorJSONResponse Error

type ListNamespacesRequestObject struct {
}

type ListNamespacesResponseObject interface {
	VisitListNamespacesResponse(w http.ResponseWriter) error
}

type ListNamespaces200JSONResponse NamespaceList

func (response ListNamespaces200JSONResponse) VisitListNamespacesResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type ListNamespaces500JSONResponse Error

func (response ListNamespaces500JSONResponse) VisitListNamespacesResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

type CreateNamespaceRequestObject struct {
	Body *CreateNamespaceJSONRequestBody
}

type CreateNamespaceResponseObject interface {
	VisitCreateNamespaceResponse(w http.ResponseWriter) error
}

type CreateNamespace201JSONResponse Namespace

func (response CreateNamespace201JSONResponse) VisitCreateNamespaceResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(201)

	return json.NewEncoder(w).Encode(response)
}

type CreateNamespace400JSONResponse struct{ BadRequestErrorJSONResponse }

func (response CreateNamespace400JSONResponse) VisitCreateNamespaceResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type CreateNamespace409JSONResponse Error

func (response CreateNamespace409JSONResponse) VisitCreateNamespaceResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(409)

	return json.NewEncoder(w).Encode(response)
}

type CreateNamespace500Response struct {
}

func (response CreateNamespace500Response) VisitCreateNamespaceResponse(w http.ResponseWriter) error {
	w.WriteHeader(500)
	return nil
}

type DeleteNamespaceRequestObject struct {
	NamespaceName NamespaceNameParam `json:"namespaceName"`
}

type DeleteNamespaceResponseObject interface {
	VisitDeleteNamespaceResponse(w http.ResponseWriter) error
}

type DeleteNamespace204Response struct {
}

func (response DeleteNamespace204Response) VisitDeleteNamespaceResponse(w http.ResponseWriter) error {
	w.WriteHeader(204)
	return nil
}

type DeleteNamespace404JSONResponse struct{ NotFoundErrorJSONResponse }

func (response DeleteNamespace404JSONResponse) VisitDeleteNamespaceResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(404)

	return json.NewEncoder(w).Encode(response)
}

type DeleteNamespace409JSONResponse Error

func (response DeleteNamespace409JSONResponse) VisitDeleteNamespaceResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(409)

	return json.NewEncoder(w).Encode(response)
}

type DeleteNamespace500Response struct {
}

func (response DeleteNamespace500Response) VisitDeleteNamespaceResponse(w http.ResponseWriter) error {
	w.WriteHeader(500)
	return nil
}

type GetNamespaceRequestObject struct {
	NamespaceName NamespaceNameParam `json:"namespaceName"`
}

type GetNamespaceResponseObject interface {
	VisitGetNamespaceResponse(w http.ResponseWriter) error
}

type GetNamespace200JSONResponse Namespace

func (response GetNamespace200JSONResponse) VisitGetNamespaceResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type GetNamespace404JSONResponse struct{ NotFoundErrorJSONResponse }

func (response GetNamespace404JSONResponse) VisitGetNamespaceResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(404)

	return json.NewEncoder(w).Encode(response)
}

type GetNamespace500Response struct {
}

func (response GetNamespace500Response) VisitGetNamespaceResponse(w http.ResponseWriter) error {
	w.WriteHeader(500)
	return nil
}

type UpdateNamespaceRequestObject struct {
	NamespaceName NamespaceNameParam `json:"namespaceName"`
	Body          *UpdateNamespaceJSONRequestBody
}

type UpdateNamespaceResponseObject interface {
	VisitUpdateNamespaceResponse(w http.ResponseWriter) error
}

type UpdateNamespace200JSONResponse Namespace

func (response UpdateNamespace200JSONResponse) VisitUpdateNamespaceResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type UpdateNamespace400JSONResponse struct{ BadRequestErrorJSONResponse }

func (response UpdateNamespace400JSONResponse) VisitUpdateNamespaceResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type UpdateNamespace404JSONResponse struct{ NotFoundErrorJSONResponse }

func (response UpdateNamespace404JSONResponse) VisitUpdateNamespaceResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(404)

	return json.NewEncoder(w).Encode(response)
}

type UpdateNamespace500Response struct {
}

func (response UpdateNamespace500Response) VisitUpdateNamespaceResponse(w http.ResponseWriter) error {
	w.WriteHeader(500)
	return nil
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// Получить список неймспейсов
	// (GET /namespaces)
	ListNamespaces(ctx context.Context, request ListNamespacesRequestObject) (ListNamespacesResponseObject, error)
	// Создать новый неймспейс
	// (POST /namespaces)
	CreateNamespace(ctx context.Context, request CreateNamespaceRequestObject) (CreateNamespaceResponseObject, error)
	// Удалить неймспейс
	// (DELETE /namespaces/{namespaceName})
	DeleteNamespace(ctx context.Context, request DeleteNamespaceRequestObject) (DeleteNamespaceResponseObject, error)
	// Получить неймспейс по его имени
	// (GET /namespaces/{namespaceName})
	GetNamespace(ctx context.Context, request GetNamespaceRequestObject) (GetNamespaceResponseObject, error)
	// Обновить неймспейс
	// (PUT /namespaces/{namespaceName})
	UpdateNamespace(ctx context.Context, request UpdateNamespaceRequestObject) (UpdateNamespaceResponseObject, error)
}

type StrictHandlerFunc = strictecho.StrictEchoHandlerFunc
type StrictMiddlewareFunc = strictecho.StrictEchoMiddlewareFunc

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
}

// ListNamespaces operation middleware
func (sh *strictHandler) ListNamespaces(ctx echo.Context) error {
	var request ListNamespacesRequestObject

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.ListNamespaces(ctx.Request().Context(), request.(ListNamespacesRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "ListNamespaces")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(ListNamespacesResponseObject); ok {
		return validResponse.VisitListNamespacesResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// CreateNamespace operation middleware
func (sh *strictHandler) CreateNamespace(ctx echo.Context) error {
	var request CreateNamespaceRequestObject

	var body CreateNamespaceJSONRequestBody
	if err := ctx.Bind(&body); err != nil {
		return err
	}
	request.Body = &body

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.CreateNamespace(ctx.Request().Context(), request.(CreateNamespaceRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "CreateNamespace")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(CreateNamespaceResponseObject); ok {
		return validResponse.VisitCreateNamespaceResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// DeleteNamespace operation middleware
func (sh *strictHandler) DeleteNamespace(ctx echo.Context, namespaceName NamespaceNameParam) error {
	var request DeleteNamespaceRequestObject

	request.NamespaceName = namespaceName

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.DeleteNamespace(ctx.Request().Context(), request.(DeleteNamespaceRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "DeleteNamespace")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(DeleteNamespaceResponseObject); ok {
		return validResponse.VisitDeleteNamespaceResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// GetNamespace operation middleware
func (sh *strictHandler) GetNamespace(ctx echo.Context, namespaceName NamespaceNameParam) error {
	var request GetNamespaceRequestObject

	request.NamespaceName = namespaceName

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.GetNamespace(ctx.Request().Context(), request.(GetNamespaceRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetNamespace")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(GetNamespaceResponseObject); ok {
		return validResponse.VisitGetNamespaceResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// UpdateNamespace operation middleware
func (sh *strictHandler) UpdateNamespace(ctx echo.Context, namespaceName NamespaceNameParam) error {
	var request UpdateNamespaceRequestObject

	request.NamespaceName = namespaceName

	var body UpdateNamespaceJSONRequestBody
	if err := ctx.Bind(&body); err != nil {
		return err
	}
	request.Body = &body

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.UpdateNamespace(ctx.Request().Context(), request.(UpdateNamespaceRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "UpdateNamespace")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(UpdateNamespaceResponseObject); ok {
		return validResponse.VisitUpdateNamespaceResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}